made helper file questions.js for prompts
